% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_phylo.R
\name{convert_phylo}
\alias{convert_phylo}
\title{Function to subset and convert phyloseq data}
\usage{
convert_phylo(data, select)
}
\arguments{
\item{data}{a \code{phyloseq} object}

\item{select}{Name of OTU or taxa to select, must match taxa name in \code{data}}
}
\value{
A \code{data.frame} object, with elements \code{W} as the observed counts, \code{M} as the sequencing depth, and the sample data with their original names.
}
\description{
Function to subset and convert phyloseq data
}
\examples{
data(soil_phylo)
convert_phylo(soil_phylo, "OTU.43")

}
