% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sumexp.R
\name{convert_sumexp}
\alias{convert_sumexp}
\title{Function to subset and convert SummarizedExperiment data}
\usage{
convert_sumexp(data, select)
}
\arguments{
\item{data}{a \code{SummarizedExperiment} object}

\item{select}{Name of OTU or taxa to select, must match taxa name in \code{data}}
}
\value{
A \code{data.frame} object, with elements \code{W} as the observed counts, \code{M} as the sequencing depth, and the sample data with their original names.
}
\description{
Function to subset and convert SummarizedExperiment data
}
