% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corncob.R
\name{corncob-package}
\alias{corncob-package}
\alias{corncob}
\title{Corncob package documentation.}
\description{
Corncob provides methods for estimating, projecting, and plotting spatio-temporal under-five mortality rates.
}
\details{
For details on the model implemented in this package, see ...

The development version of the package will be maintained on \url{https://github.com/bryandmartin/CORNCOB}.
}
