% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{coth}
\alias{coth}
\title{Hyperbolic cotangent transformation}
\usage{
coth(x)
}
\arguments{
\item{x}{data}
}
\value{
Hyperbolic cotangent transformation of x
}
\description{
Hyperbolic cotangent transformation
}
\examples{
x <- .5
coth(x)

}
