% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabin.R
\name{dbetabin}
\alias{dbetabin}
\title{Betabinomial density}
\usage{
dbetabin(theta, W, M, X, X_star, np, npstar, link, phi.link,
  logpar = TRUE)
}
\arguments{
\item{theta}{parameters}

\item{W}{absolute abundance}

\item{M}{sample size}

\item{X}{mean covariates}

\item{X_star}{overdispersion covariates}

\item{np}{number of mean parameters}

\item{npstar}{number of overdisperion parameters}

\item{link}{Link function for mean, defaults to "logit"}

\item{phi.link}{Link function for overdispersion, defaults to "fishZ"}

\item{logpar}{Indicator of log-likelihood, defaults to TRUE}
}
\value{
Beta-binomial log-likelihood
}
\description{
Betabinomial density
}
\examples{
\dontrun{
TODO
}

}
