% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabin.R
\name{dbetabin_i}
\alias{dbetabin_i}
\title{Betabinomial density for a single sample}
\usage{
dbetabin_i(a1, a2, W, M, logpar = TRUE)
}
\description{
Betabinomial density for a single sample
}
\keyword{internal}
