% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom_ee.R
\name{dbetabinom_cts}
\alias{dbetabinom_cts}
\title{Densities of beta binomial distributions, permitting non integer x and size}
\usage{
dbetabinom_cts(x, size, prob, rho = 0, log = FALSE)
}
\arguments{
\item{x}{the value at which defined the density}

\item{size}{number of trials}

\item{prob}{the probability of success}

\item{rho}{the correlation parameter}

\item{log}{if TRUE, log-densities p are given}
}
\description{
In some cases we may not have integer W and M's. In these cases,
we can still use corncob to estimate parameters, but we need to think of them as
no longer coming from the specific beta binomial parametric model,
and instead from an estimating equations framework.
}
\author{
Thomas W Yee

Xiangjie Xue

Amy D Willis
}
