% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBoot.R
\name{doBoot}
\alias{doBoot}
\title{Function to run a bootstrap iteration}
\usage{
doBoot(mod, mod_null, test, robust = FALSE)
}
\arguments{
\item{mod}{an object of class \code{bbdml}}

\item{mod_null}{an object of class \code{bbdml}}

\item{test}{Character. Hypothesis testing procedure to use. One of \code{"Wald"} or \code{"LRT"} (likelihood ratio test).}

\item{robust}{Should robust standard errors be used? If not, model-based standard arras are used. Logical, defaults to \code{FALSE}.}
}
\value{
test statistic from one bootstrap iteration
}
\description{
Internal function. Not intended for users.
}
