% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{fishZ}
\alias{fishZ}
\title{Fisher's z transformation}
\usage{
fishZ(x)
}
\arguments{
\item{x}{data}
}
\value{
Fisher's z transformation of x
}
\description{
Fisher's z transformation
}
\examples{
x <- .5
fishZ(x)

}
