% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_betast.R
\name{gr_betast}
\alias{gr_betast}
\title{Overdispersion Parameter Gradient Vector}
\usage{
gr_betast(theta, W, M, X, X_star, np, npstar, logpar = TRUE)
}
\arguments{
\item{theta}{parameters}

\item{W}{absolute abundance}

\item{M}{sample size}

\item{X}{mean covariates}

\item{X_star}{overdispersion covariates}

\item{np}{number of mean parameters}

\item{npstar}{number of overdisperion parameters}

\item{logpar}{Indicator of log-likelihood, defaults to TRUE}
}
\value{
Gradient of likelihood with respect to overdispersion parameters
}
\description{
Overdispersion Parameter Gradient Vector
}
\examples{
\dontrun{
TODO
}

}
