% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_betast.R
\name{gr_betast_i}
\alias{gr_betast_i}
\title{Overdispersion Gradient for a single sample}
\usage{
gr_betast_i(k, k_star, W, M, coth_st)
}
\description{
Overdispersion Gradient for a single sample
}
\keyword{internal}
