% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_full.R
\name{gr_full}
\alias{gr_full}
\title{Full gradient}
\usage{
gr_full(...)
}
\arguments{
\item{...}{See \code{\link{gr_beta}} or \code{\link{gr_betast}}}
}
\value{
Gradient of likelihood with respect to all parameters, as vector
}
\description{
Full gradient
}
\examples{
\dontrun{
TODO
}

}
