% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{hessian}
\alias{hessian}
\title{Compute Hessian}
\usage{
hessian(mod, numerical = FALSE)
}
\arguments{
\item{mod}{model fit from bbdml}

\item{numerical}{Boolean numerical Hessian. Not as stable. Defaults to FALSE}
}
\value{
Analytic Hessian
}
\description{
Compute Hessian
}
\examples{
\dontrun{
TODO
}

}
