% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDbetabinom.R
\name{HPDbetabinom}
\alias{HPDbetabinom}
\title{Get HPD interval of beta binom}
\usage{
HPDbetabinom(percent, size, mu, phi)
}
\arguments{
\item{percent}{percent interval}

\item{size}{size}

\item{mu}{probability parameter}

\item{phi}{overdispersion parameter}
}
\value{
quantile
}
\description{
Get HPD interval of beta binom
}
\examples{
\dontrun{
TODO
}
}
