% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_phylo.R
\docType{data}
\name{ibd_phylo}
\alias{ibd_phylo}
\title{IBD data.}
\format{A phyloseq-class experiment-level object with an OTU table and sample data.}
\usage{
ibd_phylo
}
\description{
A phyloseq object with an OTU table and sample data from an IBD microbiome study.
}
\references{
Duvallet, C., Gibbons, S. M., Gurry, T., Irizarry, R. A., & Alm, E. J. (2017). \emph{Meta-analysis of gut microbiome studies identifies disease-specific and shared responses}. Nature communications, 8(1), 1784. <doi.org/10.1038/s41467-017-01973-8>.

Duvallet, C., Gibbons, S., Gurry, T., Irizarry, R., & Alm, E. (2017). \emph{MicrobiomeHD: the human gut microbiome in health and disease [Data set]}. Zenodo. <doi.org/10.5281/zenodo.1146764>.
}
\keyword{datasets}
