% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_phylo_otu.R
\docType{data}
\name{ibd_phylo_otu}
\alias{ibd_phylo_otu}
\title{IBD data, OTU count data frame.}
\format{
A data frame of OTU counts.
}
\usage{
ibd_phylo_otu
}
\description{
OTU data frame from a phyloseq object from an IBD microbiome study.
}
\references{
Papa, E., Docktor, M., Smillie, C., Weber, S., Preheim, S. P., Gevers, D., Giannoukos, G., Ciulla, D., Tabbaa, D., Ingram, J., Schauer, D. B., Ward, D. V., Korzenik, J. R., Xavier, R. J., Bousvaros, A., Alm, E. J. & Schauer, D. B. (2012). \emph{Non-invasive mapping of the gastrointestinal microbiota identifies children with inflammatory bowel disease}. PloS One, 7(6), e39242. <doi.org/10.1371/journal.pone.0039242>.

Duvallet, C., Gibbons, S., Gurry, T., Irizarry, R., & Alm, E. (2017). \emph{MicrobiomeHD: the human gut microbiome in health and disease [Data set]}. Zenodo. <doi.org/10.5281/zenodo.1146764>.
}
\keyword{datasets}
