% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{invfishZ}
\alias{invfishZ}
\title{Inverse Fisher's z transformation}
\usage{
invfishZ(x)
}
\arguments{
\item{x}{data}
}
\value{
Inverse Fisher's z transformation of x
}
\description{
Inverse Fisher's z transformation
}
\examples{
x <- .5
invfishZ(x)

}
