% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{data}
}
\value{
Inverse logit transformation of x
}
\description{
Inverse logit transformation
}
\examples{
x <- .5
invlogit(x)

}
