% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbLRT.R
\name{pbLRT}
\alias{pbLRT}
\title{Parametric bootstrap likelihood ratio test}
\usage{
pbLRT(mod, mod_null, B = 1000)
}
\arguments{
\item{mod}{Unrestricted model fit from \code{bbdml}.}

\item{mod_null}{Restricted model fit from \code{bbdml}.}

\item{B}{Integer. Defaults to \code{1000}. Number of bootstrap iterations.}
}
\value{
P-value.
}
\description{
Parametric bootstrap likelihood ratio test
}
\examples{
\dontrun{
TODO
}
}
