% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differentialTest.R
\name{plot.differentialTest}
\alias{plot.differentialTest}
\title{differentialTest plot function}
\usage{
\method{plot}{differentialTest}(x, level = NULL, data_only = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{differentialTest}}

\item{level}{(Optional). Character vector. Desired taxonomic levels for taxa labels.}

\item{data_only}{(Optional). Default \code{FALSE}. Boolean. If \code{TRUE}, only returns data frame.}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
Object of class \code{ggplot}. Plot of coefficients from models for significant taxa from \code{differentialTest}
}
\description{
differentialTest plot function
}
\examples{
\dontrun{
# phyloseq example
data(soil_phylo)
soil <- soil_phylo \%>\%
phyloseq::subset_samples(DayAmdmt \%in\% c(11,21)) \%>\%
phyloseq::tax_glom("Phylum")
da_analysis <- differentialTest(formula = ~ DayAmdmt,
                                phi.formula = ~ DayAmdmt,
                                formula_null = ~ 1,
                                phi.formula_null = ~ DayAmdmt,
                                test = "Wald", boot = FALSE,
                                data = soil,
                                fdr_cutoff = 0.05)
plot(da_analysis, level = "Phylum")
}

}
