% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_bbdml.R
\name{print.bbdml}
\alias{print.bbdml}
\title{Print function}
\usage{
\method{print}{bbdml}(x, digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{Object of class \code{bbdml}}

\item{digits}{the nnumber of significant digits to use when printing.}

\item{signif.stars}{logical. If \code{TRUE}, `significance stars' are printed for each coefficient.}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
Model summary
}
\description{
Print function
}
\examples{
\dontrun{
TODO
}
}
