% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_differentialTest.R
\name{print.differentialTest}
\alias{print.differentialTest}
\title{differentialTest print function}
\usage{
\method{print}{differentialTest}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bbdml}}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
\code{NULL}. Displays printed \code{differentialTest} summary.
}
\description{
differentialTest print function
}
\examples{
\dontrun{
# phyloseq example
data(soil_phylo)
soil <- soil_phylo \%>\%
phyloseq::subset_samples(DayAmdmt \%in\% c(11,21)) \%>\%
phyloseq::tax_glom("Phylum")
da_analysis <- differentialTest(formula = ~ DayAmdmt,
                                phi.formula = ~ DayAmdmt,
                                formula_null = ~ 1,
                                phi.formula_null = ~ DayAmdmt,
                                test = "Wald", boot = FALSE,
                                data = soil,
                                fdr_cutoff = 0.05)
print(da_analysis)
}

}
