% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_bbdml.R
\name{print.summary.bbdml}
\alias{print.summary.bbdml}
\title{Print summary function}
\usage{
\method{print}{summary.bbdml}(x, digits = max(3L, getOption("digits") -
  3L), signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{Model summary output from \code{\link{summary.bbdml}}}

\item{digits}{the nnumber of significant digits to use when printing.}

\item{signif.stars}{logical. If \code{TRUE}, `significance stars' are printed for each coefficient.}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
Model summary print
}
\description{
Print summary function
}
\examples{
\dontrun{
TODO
}
}
