% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbetabinom.R
\name{qbetabinom}
\alias{qbetabinom}
\title{Get quantiles of beta binom}
\usage{
qbetabinom(p, size, mu, phi)
}
\arguments{
\item{p}{probability for quantile}

\item{size}{size}

\item{mu}{probability parameter}

\item{rho}{overdispersion parameter}
}
\value{
quantile
}
\description{
Get quantiles of beta binom
}
\examples{
\dontrun{
TODO
}
}
