% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandSE.R
\name{sandSE}
\alias{sandSE}
\title{Compute sandwich standard error}
\usage{
sandSE(mod, numerical = FALSE)
}
\arguments{
\item{mod}{model fit from bbdml}

\item{numerical}{Boolean indicator for numerical score and hessian. Not as stable. Defaults to FALSE}
}
\value{
Sandwich variance-covariance matrix
}
\description{
Compute sandwich standard error
}
\examples{
\dontrun{
TODO
}

}
