% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.bbdml}
\alias{simulate.bbdml}
\title{Simulate from BBD model fit}
\usage{
\method{simulate}{bbdml}(object, nsim, seed = NULL, ...)
}
\arguments{
\item{object}{model output from \code{\link{bbdml}}}

\item{nsim}{number of simulations}

\item{seed}{object specifying random seed if desired}

\item{...}{optional additional arguments}
}
\value{
nsim simulations from mod
}
\description{
Simulate from BBD model fit
}
