% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_phylo.R
\docType{data}
\name{soil_phylo}
\alias{soil_phylo}
\title{Soil data.}
\format{A phyloseq-class experiment-level object with an OTU table and sample data.}
\usage{
soil_phylo
}
\description{
A phyloseq object with an OTU table and sample data from a soil microbiome study.
}
\references{
Whitman, T., Pepe-Ranney, C., Enders, A., Koechli, C., Campbell, A.,  Buckley, D. H., Lehmann, J. (2016). \emph{Dynamics of microbial community composi-tion and soil organic carbon mineralization in soil following addition of pyrogenic andfresh organic matter}. The ISME journal, 10(12):2918.
}
\keyword{datasets}
