% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_phylo_otu.R
\docType{data}
\name{soil_phylo_otu}
\alias{soil_phylo_otu}
\title{Soil data, otu table as data frame.}
\format{
A phyloseq-class experiment-level object with an OTU table.
\describe{
\item{otu_table}{OTU table with 7,770 taxa and 119 samples}
}
}
\usage{
soil_phylo_otu
}
\description{
A data frame made from a soil `phyloseq` object with only otu count data.
}
\references{
Whitman, T., Pepe-Ranney, C., Enders, A., Koechli, C., Campbell, A.,  Buckley, D. H., Lehmann, J. (2016). \emph{Dynamics of microbial community composi-tion and soil organic carbon mineralization in soil following addition of pyrogenic andfresh organic matter}. The ISME journal, 10(12):2918. <doi: 10.1038/ismej.2016.68>.
}
\keyword{datasets}
