% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_phylum_small_otu.R
\docType{data}
\name{soil_phylum_small_otu}
\alias{soil_phylum_small_otu}
\title{Small soil phylum data for examples, otu table as data frame}
\format{
A phyloseq-class experiment-level object with an OTU table.
\describe{
\item{otu_table}{OTU table with 39 taxa and 32 samples}
}
}
\usage{
soil_phylum_small_otu
}
\description{
A small subset of \code{\link{soil_phylo}} used for examples. A data frame made from the `phyloseq` object with only otu counts.
}
\references{
Whitman, T., Pepe-Ranney, C., Enders, A., Koechli, C., Campbell, A.,  Buckley, D. H., Lehmann, J. (2016). \emph{Dynamics of microbial community composi-tion and soil organic carbon mineralization in soil following addition of pyrogenic andfresh organic matter}. The ISME journal, 10(12):2918. <doi: 10.1038/ismej.2016.68>.
}
\keyword{datasets}
