% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bbdml.R
\name{summary.bbdml}
\alias{summary.bbdml}
\title{Summary function}
\usage{
\method{summary}{bbdml}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bbdml}}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
Model summary
}
\description{
Summary function
}
\examples{
\dontrun{
TODO
}
}
