% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{waldchisq_test}
\alias{waldchisq_test}
\title{Wald-type chi-squared test statistic}
\usage{
waldchisq_test(mod, restrictions, testonly = NULL)
}
\arguments{
\item{mod}{unrestricted model fit from \code{bbdml}}

\item{restrictions}{Numeric vector indicating the parameters to test, or character vector with name of variable to test.}

\item{testonly}{Optional. If set to \code{"mu"}, then only the effect of the variable provided in \code{restrictions} on the mean is tested. If set to \code{"phi"}, then only the effect of the variable provided in \code{restrictions} on the variance is tested. Otherwise, the effects are jointly tested.}
}
\value{
Test statistic.
}
\description{
Wald-type chi-squared test statistic
}
