% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{waldt}
\alias{waldt}
\title{Wald-type t test}
\usage{
waldt(mod)
}
\arguments{
\item{mod}{model fit from \code{bbdml}}
}
\value{
Matrix with wald test statistics and p-values. Only performs univariate tests
}
\description{
Wald-type t test
}
\examples{
\dontrun{
TODO
}
}
