% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{waldtest}
\alias{waldtest}
\title{Wald test}
\usage{
waldtest(mod)
}
\arguments{
\item{mod}{model fit from bbdml}
}
\value{
Matrix with wald test statistics and p-values
}
\description{
Wald test
}
\examples{
\dontrun{
TODO
}
}
