% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn_phyloseq.R
\name{warn_phyloseq}
\alias{warn_phyloseq}
\title{Function to throw error if the `phyloseq` package is called but it is not installed}
\usage{
warn_phyloseq()
}
\description{
Function to throw error if the `phyloseq` package is called but it is not installed
}
