% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn_phyloseq.R
\name{warn_sumexp}
\alias{warn_sumexp}
\title{Function to throw error if the `SummarizedExperiment` package is called but it is not installed}
\usage{
warn_sumexp()
}
\description{
Function to throw error if the `SummarizedExperiment` package is called but it is not installed
}
