% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbdml.R
\name{bbdml}
\alias{bbdml}
\title{Maximum Likelihood for the Beta-binomial Distribution}
\usage{
bbdml(formula, phi.formula, data, link = "logit", phi.link = "logit",
  method = "trust", control = list(maxit = 1000, reltol = 1e-14),
  numerical = FALSE, nstart = 1, inits = NULL, ...)
}
\arguments{
\item{formula}{Formula for mean}

\item{phi.formula}{Formula for overdispersion}

\item{data}{Data frame or \code{phyloseq} object}

\item{link}{Link function for mean, defaults to "logit"}

\item{phi.link}{Link function for overdispersion, defaults to "logit"}

\item{method}{Method for optimization (see \code{\link{optimr}})}

\item{control}{Optimization control parameters (see \code{\link{optimr}})}

\item{numerical}{Indicator to use numerical derivative, useful for testing}

\item{nstart}{Number of starts for optimization, defaults to 1}

\item{inits}{Optional argument to specify initializations, defaults to NULL}

\item{...}{Additional arguments for \code{\link{optimr}}}
}
\value{
BBD Model fit
}
\description{
Maximum Likelihood for the Beta-binomial Distribution
}
