% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbdml.R
\name{bbdml}
\alias{bbdml}
\title{Maximum Likelihood for the Beta-binomial Distribution}
\usage{
bbdml(formula, phi.formula, data, link = "logit", phi.link = "fishZ",
  phi.init = NULL, method = "L-BFGS-B", control = list(maxit = 1000,
  reltol = 1e-14, factr = 1000), numerical = FALSE, ...)
}
\arguments{
\item{formula}{Formula for mean}

\item{phi.formula}{Formula for overdispersion}

\item{data}{Data frame}

\item{link}{Link function for mean, defaults to "logit"}

\item{phi.link}{Link function for overdispersion, defaults to "fishZ"}

\item{phi.init}{Optional starting value for overdispersion}

\item{method}{Method for optimization (see \code{\link{optimr}})}

\item{control}{Optimization control parameters (see \code{\link{optimr}})}

\item{numerical}{Indicator to use numerical derivative, useful for testing}

\item{...}{Additional arguments for \code{\link{optimr}}}
}
\value{
BBD Model fit
}
\description{
Maximum Likelihood for the Beta-binomial Distribution
}
