% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNested.R
\name{checkNested}
\alias{checkNested}
\title{Check for nested models}
\usage{
checkNested(mod, mod_null)
}
\arguments{
\item{mod}{an object of class \code{bbdml}}

\item{mod_null}{an object of class \code{bbdml}}
}
\value{
\code{TRUE} if \code{mod_null} is nested within \code{mod}, otherwise it throws an error.
}
\description{
Check for nested models
}
\examples{
data(soil_phylum_small_otu1)
mod1 <- bbdml(formula = cbind(W, M - W) ~ DayAmdmt,
phi.formula = ~ DayAmdmt,
data = soil_phylum_small_otu1)

mod2 <- bbdml(formula = cbind(W, M - W) ~ 1,
phi.formula = ~ 1,
data = soil_phylum_small_otu1)

checkNested(mod1, mod2)

}
