% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNested.R
\name{checkNested}
\alias{checkNested}
\title{Check for nested models}
\usage{
checkNested(mod, mod_null)
}
\arguments{
\item{mod}{model fit from bbdml}

\item{mod_null}{model fit from bbdml}
}
\value{
Boolean. \code{TRUE} if \code{mod_null} is nested within \code{mod}.
}
\description{
Check for nested models
}
\examples{
\dontrun{
TODO
}
}
