% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_phylo.R
\name{convert_phylo}
\alias{convert_phylo}
\title{Function to subset and convert phyloseq data}
\usage{
convert_phylo(data, select)
}
\arguments{
\item{data}{\code{phyloseq} object}

\item{select}{Name of OTU or taxa to select, must match taxa name in \code{data}}
}
\value{
A data.frame object
}
\description{
Function to subset and convert phyloseq data
}
\examples{
\dontrun{
data(soil_phylo)
convert_phylo(soil_phylo, "OTU1")
}

}
