% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corncob.R
\name{corncob-package}
\alias{corncob-package}
\alias{corncob}
\title{Corncob package documentation.}
\value{
No return value. Created for documentation.
}
\description{
Corncob provides methods for estimating and plotting count data. Specifically, corncob is designed to account for the challenges of modeling sequencing data from microbial abundance studies.
}
\details{
For details on the model implemented in this package, see Martin et al. (2020) <doi:10.1214/19-AOAS1283>.

The development version of the package will be maintained on \url{https://github.com/bryandmartin/corncob}.
}
