% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialTest.R
\name{differentialTest}
\alias{differentialTest}
\title{Identify differentially-abundant and differentially-variable taxa}
\usage{
differentialTest(formula, phi.formula, formula_null, phi.formula_null, data,
  link = "logit", phi.link = "logit", fdr_cutoff = 0.05, fdr = "fdr",
  ...)
}
\arguments{
\item{formula}{Formula for mean, without response}

\item{phi.formula}{Formula for overdispersion, without response}

\item{formula_null}{Formula for mean under null, without response}

\item{phi.formula_null}{Formula for overdispersion under null, without response}

\item{data}{Data frame or \code{phyloseq} object}

\item{link}{Link function for mean, defaults to "logit"}

\item{phi.link}{Link function for overdispersion, defaults to "logit"}

\item{fdr_cutoff}{Desired type 1 error rate}

\item{fdr}{False discovery rate control method, defaults to "fdr"}

\item{...}{Additional arguments for \code{\link{bbdml}}}
}
\value{
List of differentially-abundant taxa and differentially-variable taxa
}
\description{
Identify differentially-abundant and differentially-variable taxa
}
