% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBoot.R
\name{doBoot}
\alias{doBoot}
\title{Function to run a bootstrap iteration}
\usage{
doBoot(mod, mod_null, test)
}
\arguments{
\item{mod}{Unrestricted model fit from \code{bbdml}.}

\item{mod_null}{Restricted model fit from \code{bbdml}.}

\item{test}{Character string. Type of test, either "LRT" or "Wald".}
}
\value{
test statistic from one bootstrap iteration
}
\description{
Function to run a bootstrap iteration
}
