% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInits.R
\name{genInits}
\alias{genInits}
\title{Generate initialization for optimization}
\usage{
genInits(nstart, max.call = NULL, temperature = NULL, W, M, X, X_star,
  np, npstar, link, phi.link, logpar, lower = NULL, upper = NULL)
}
\arguments{
\item{nstart}{Number of starts for optimization}

\item{max.call}{TODO}

\item{temperature}{TODO}

\item{W}{absolute abundance}

\item{M}{sample size}

\item{X}{mean covariates}

\item{X_star}{overdispersion covariates}

\item{np}{number of mean parameters}

\item{npstar}{number of overdisperion parameters}

\item{link}{Link function for mean}

\item{phi.link}{Link function for overdispersion}

\item{logpar}{Indicator of log-likelihood}

\item{lower}{Lower bound for parameter values. Defaults to -20}

\item{upper}{Upper bound for parameter values. Defaults to 20.}
}
\value{
Matrix of initializations
}
\description{
Generate initialization for optimization
}
\examples{
\dontrun{
TODO
}
}
