% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRestrictionTerms.R
\name{getRestrictionTerms}
\alias{getRestrictionTerms}
\title{Get index of restricted terms for Wald test}
\usage{
getRestrictionTerms(mod, mod_null)
}
\arguments{
\item{mod}{unrestricted model fit from \code{bbdml}}

\item{mod_null}{restricted model fit from \code{bbdml}}
}
\value{
Numeric vector of index of restricted terms.
}
\description{
Get index of restricted terms for Wald test
}
