% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_full.R
\name{gr_full}
\alias{gr_full}
\title{Parameter Gradient Vector}
\usage{
gr_full(theta, W, M, X, X_star, np, npstar, link, phi.link,
  logpar = TRUE)
}
\arguments{
\item{theta}{parameters}

\item{W}{absolute abundance}

\item{M}{sample size}

\item{X}{mean covariates}

\item{X_star}{overdispersion covariates}

\item{np}{number of mean parameters}

\item{npstar}{number of overdisperion parameters}

\item{link}{Link function for mean, defaults to "logit"}

\item{phi.link}{Link function for overdispersion, defaults to "fishZ"}

\item{logpar}{Indicator of log-likelihood, defaults to TRUE}
}
\value{
Gradient of likelihood with respect to parameters
}
\description{
Parameter Gradient Vector
}
\examples{
\dontrun{
TODO
}

}
