% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{lrtest}
\alias{lrtest}
\title{Likelihood ratio test}
\usage{
lrtest(mod1, mod2)
}
\arguments{
\item{mod1}{model fit from bbdml, nested within \code{mod2}}

\item{mod2}{model fit from bbdml}
}
\value{
Matrix with likelihood ratio test statistic and p-value
}
\description{
Likelihood ratio test
}
\examples{
\dontrun{
TODO
}
}
