% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{lrtest}
\alias{lrtest}
\title{Likelihood ratio test}
\usage{
lrtest(mod, mod_null)
}
\arguments{
\item{mod}{model fit from bbdml}

\item{mod_null}{model fit from bbdml, nested within \code{mod}}
}
\value{
P-value.
}
\description{
Likelihood ratio test
}
\examples{
\dontrun{
TODO
}
}
