% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objfun.R
\name{objfun}
\alias{objfun}
\title{Objective function}
\usage{
objfun(theta, W, M, X, X_star, np, npstar, link, phi.link)
}
\arguments{
\item{theta}{Numeric vector. Parameters associated with \code{X} and \code{X_star}}

\item{W}{Numeric vector of counts}

\item{M}{Numeric vector of sequencing depth}

\item{X}{Matrix of covariates associated with abundance (including intercept)}

\item{X_star}{Matrix of covariates associated with dispersion (including intercept)}

\item{np}{Number of covariates associated with abundance (including intercept)}

\item{npstar}{Number of covariates associated with dispersion (including intercept)}

\item{link}{ink function for abundance covariates}

\item{phi.link}{ink function for dispersion covariates}
}
\value{
List of negative log-likelihood, gradient, and hessian
}
\description{
Used for internal optimization. Not intended for users.
}
