% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_to_taxonomy.R
\name{otu_to_taxonomy}
\alias{otu_to_taxonomy}
\title{Transform OTUs to their taxonomic label}
\usage{
otu_to_taxonomy(OTU, data)
}
\arguments{
\item{OTU}{String vector. Names of OTU labels in \code{data}}

\item{data}{\code{phyloseq} object with a taxonomy table}
}
\value{
String vector. Names of taxonomic labels matching labels of \code{OTU}.
}
\description{
Transform OTUs to their taxonomic label
}
\examples{
\dontrun{
TODO
}

}
