% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_to_taxonomy.R
\name{otu_to_taxonomy}
\alias{otu_to_taxonomy}
\title{Transform OTUs to their taxonomic label}
\usage{
otu_to_taxonomy(OTU, data)
}
\arguments{
\item{OTU}{String vector. Names of OTU labels in \code{data}}

\item{data}{\code{phyloseq} object with a taxonomy table}
}
\value{
String vector. Names of taxonomic labels matching labels of \code{OTU}.
}
\description{
Transform OTUs to their taxonomic label
}
\examples{
\dontrun{
data(soil_phylo)
soil <- soil_phylo \%>\%
phyloseq::subset_samples(DayAmdmt \%in\% c(11,21)) \%>\%
phyloseq::tax_glom("Phylum")
da_analysis <- differentialTest(formula = ~ DayAmdmt,
                                phi.formula = ~ DayAmdmt,
                                formula_null = ~ 1,
                                phi.formula_null = ~ DayAmdmt,
                                test = "Wald", boot = FALSE,
                                data = soil,
                                fdr_cutoff = 0.05)
otu_to_taxonomy(OTU = da_analysis$significant_taxa, data = soil)
}

}
