% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bbdml.R
\name{plot.bbdml}
\alias{plot.bbdml}
\title{Plotting function (in beta)}
\usage{
\method{plot}{bbdml}(mod, RA = FALSE)
}
\arguments{
\item{mod}{Output from \code{\link{bbdml}}}

\item{RA}{Boolean indicator for whether to plot on relative abundance scale. Defaults to FALSE.}
}
\value{
Model plot
}
\description{
Plotting function (in beta)
}
\examples{
\dontrun{
TODO
}
}
