% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bbdml.R
\name{plot.bbdml}
\alias{plot.bbdml}
\title{Plotting function (in beta)}
\usage{
\method{plot}{bbdml}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{bbdml}}}

\item{...}{See details}
}
\value{
Model plot
}
\description{
Plotting function (in beta)
}
\details{
... can include boolean indicator called RA for whether to plot on relative abundance scale, which defaults to FALSE. It can also include "group", a string of a covariate in the model data, if you want the plot to be colored by group.
}
\examples{
\dontrun{
TODO
}
}
