% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bbdml.R
\name{plot.bbdml}
\alias{plot.bbdml}
\title{Plotting function}
\usage{
\method{plot}{bbdml}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{bbdml}}}

\item{...}{See details}
}
\value{
Model plot
}
\description{
Plotting function
}
\details{
... can include boolean indicator called AA for whether to plot on absolute abundance scale, which defaults to FALSE. It can also include "color", a string of a covariate in the model data, if you want the plot to be colored by color.
}
\examples{
\dontrun{
TODO
}
}
