% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bbdml.R
\name{plot.bbdml}
\alias{plot.bbdml}
\title{Plotting function}
\usage{
\method{plot}{bbdml}(x, AA = FALSE, color = NULL, shape = NULL,
  facet = NULL, title = NULL, ...)
}
\arguments{
\item{x}{Output from \code{\link{bbdml}} of class \code{bbdml}.}

\item{AA}{(Optional). Default \code{FALSE}. Boolean indicator for whether to plot on absolute abundance scale}

\item{color}{(Optional). Default \code{NULL}. The sample variable to map to different colors. Can be a single character string of the variable name in \code{sample_data} or a custom supplied vector with length equal to the number of samples.}

\item{shape}{(Optional). Default \code{NULL}. The sample variable to map to different shapes. Can be a single character string of the variable name in \code{sample_data} or a custom supplied vector with length equal to the number of samples.}

\item{facet}{(Optional). Default \code{NULL}. The sample variable to map to different panels in a facet grid. Must be a single character string of a variable name in \code{sample_data}.}

\item{title}{(Optional). Default NULL. Character string. The main title for the graphic.}

\item{...}{There are no optional parameters at this time.}
}
\value{
Model plot
}
\description{
Plotting function
}
\examples{
\dontrun{
TODO
}
}
