% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_differentialTest.R
\name{print.differentialTest}
\alias{print.differentialTest}
\title{differentialTest print function}
\usage{
\method{print}{differentialTest}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bbdml}}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
\code{NULL}. Displays printed \code{differentialTest} summary.
}
\description{
differentialTest print function
}
\examples{
# phyloseq example
data(soil_phylum_small_sample)
data(soil_phylum_small_otu)
da_analysis <- differentialTest(formula = ~ DayAmdmt,
                                phi.formula = ~ DayAmdmt,
                                formula_null = ~ 1,
                                phi.formula_null = ~ DayAmdmt,
                                test = "Wald", boot = FALSE,
                                data = soil_phylum_small_otu,
                                sample_data = soil_phylum_small_sample,
                                fdr_cutoff = 0.05,
                                try_only = 1:5)
print(da_analysis)

}
