% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_bbdml.R
\name{print.summary.bbdml}
\alias{print.summary.bbdml}
\title{Print ummary function}
\usage{
\method{print}{summary.bbdml}(x, digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{...}{See details}

\item{object}{Model summary output from \code{\link{summary.bbdml}}}
}
\value{
Model summary print
}
\description{
Print ummary function
}
\details{
... TODO
}
\examples{
\dontrun{
TODO
}
}
