% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandSE.R
\name{sandSE}
\alias{sandSE}
\title{Compute sandwich standard errors. Legacy function. Use sand_vcov instead.}
\usage{
sandSE(mod, numerical = FALSE)
}
\arguments{
\item{mod}{an object of class \code{bbdml}}

\item{numerical}{Boolean. Defaults to \code{FALSE}. Indicator of whether to use the numeric Hessian and score (not recommended).}
}
\value{
Sandwich variance-covariance matrix
}
\description{
Compute sandwich standard errors. Legacy function. Use sand_vcov instead.
}
\examples{
data(soil_phylum_small_otu1)
mod <- bbdml(formula = cbind(W, M - W) ~ DayAmdmt,
phi.formula = ~ DayAmdmt,
data = soil_phylum_small_otu1)
sandSE(mod)

}
