% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandSE.R
\name{sandSE}
\alias{sandSE}
\title{Compute sandwich standard error}
\usage{
sandSE(mod, numerical = FALSE)
}
\arguments{
\item{mod}{an object of class \code{bbdml}}

\item{numerical}{Boolean. Defaults to \code{FALSE}. Indicator of whether to use the numeric Hessian and score (not recommended).}
}
\value{
Sandwich variance-covariance matrix
}
\description{
Compute sandwich standard error
}
\examples{
data(soil_phylo)
soil <- soil_phylo \%>\%
phyloseq::subset_samples(DayAmdmt \%in\% c(11,21)) \%>\%
phyloseq::tax_glom("Phylum")
mod <- bbdml(formula = OTU.1 ~ DayAmdmt,
phi.formula = ~ DayAmdmt,
data = soil)
sandSE(mod)

}
