% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Compute score}
\usage{
score(mod, numerical = FALSE, forHess = FALSE)
}
\arguments{
\item{mod}{model fit from bbdml}

\item{numerical}{Boolean numerical score. Not as stable. Defaults to FALSE}

\item{forHess}{Boolean for whether to use to approximate Hessian. Defaults to FALSE.}
}
\value{
Analytic score
}
\description{
Compute score
}
\examples{
\dontrun{
TODO
}

}
