% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Compute score at the MLE}
\usage{
score(mod, numerical = FALSE, get_score_covariance = FALSE)
}
\arguments{
\item{mod}{an object of class \code{bbdml}}

\item{numerical}{Boolean. Defaults to \code{FALSE}. Indicator of whether to use the numeric Hessian and score (not recommended).}

\item{get_score_covariance}{Boolean. Defaults to \code{FALSE}. Should we return a robust estimate of variance of score: $hat{B}(hat{theta}) = sum_i G(hat{theta}; W_i) G(hat{theta}; W_i)^T $. This parameter is not intended for users.}
}
\value{
Score at the MLE. For $G(theta, w)$ score function, returns $sum_i G(hat{theta}, W_i)$ if get_score_covariance = FALSE.
}
\description{
Compute score at the MLE
}
\examples{
data(soil_phylum_small)
mod <- bbdml(formula = OTU.1 ~ DayAmdmt,
phi.formula = ~ DayAmdmt,
data = soil_phylum_small)
score(mod)

}
