% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.bbdml}
\alias{simulate.bbdml}
\title{Simulate from beta-binomial model}
\usage{
\method{simulate}{bbdml}(object, nsim, seed = NULL, ...)
}
\arguments{
\item{nsim}{Integer. Number of simulations}

\item{seed}{Optional integer to set a random seed}

\item{...}{There are no additional parameters at this time.}

\item{mod}{an object of class \code{bbdml}}
}
\value{
\code{nsim} simulations from \code{mod}
}
\description{
Simulate from beta-binomial model
}
