% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bbdml.R
\name{summary.bbdml}
\alias{summary.bbdml}
\title{Summary function}
\usage{
\method{summary}{bbdml}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bbdml}}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
Object of class \code{summary.bbdml}. Displays printed model summary.
}
\description{
Summary function
}
\examples{
\dontrun{
data(soil_phylo)
soil <- soil_phylo \%>\%
phyloseq::subset_samples(DayAmdmt \%in\% c(11,21)) \%>\%
phyloseq::tax_glom("Phylum")
mod <- bbdml(formula = OTU.1 ~ DayAmdmt,
phi.formula = ~ DayAmdmt,
data = soil)
summary(mod)
}
}
