% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bbdml.R
\name{summary.bbdml}
\alias{summary.bbdml}
\title{Summary function}
\usage{
\method{summary}{bbdml}(object, ...)
}
\arguments{
\item{object}{Model fit output from \code{\link{bbdml}}}

\item{...}{See details}
}
\value{
Model summary
}
\description{
Summary function
}
\details{
... TODO
}
\examples{
\dontrun{
TODO
}
}
