% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{waldchisq}
\alias{waldchisq}
\title{Wald-type chi-squared test}
\usage{
waldchisq(mod, mod_null = NULL, restrictions = NULL, testonly = NULL)
}
\arguments{
\item{mod}{unrestricted model fit from \code{bbdml}}

\item{mod_null}{Optional. Restricted model fit from \code{bbdml}. If not included, need to include \code{restrictions}.}

\item{restrictions}{Optional. Numeric vector indicating the parameters to test, or character vector with name of variable to test.}

\item{testonly}{Optional. Defaults to \code{NULL}. If set to \code{"mu"}, then only the effect of the variable provided in \code{restrictions} on the mean is tested. If set to \code{"phi"}, then only the effect of the variable provided in \code{restrictions} on the variance is tested. Otherwise, the effects are jointly tested.}
}
\value{
P-value.
}
\description{
Wald-type chi-squared test
}
\examples{
\dontrun{
TODO
}
}
