% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{waldchisq}
\alias{waldchisq}
\title{Wald-type chi-squared test}
\usage{
waldchisq(mod, restrictions)
}
\arguments{
\item{mod}{unrestricted model fit from \code{bbdml}}

\item{restrictions}{Numeric vector indicating the parameters to test}
}
\value{
Matrix with wald test statistics and p-values
}
\description{
Wald-type chi-squared test
}
\examples{
\dontrun{
TODO
}
}
