% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{waldchisq}
\alias{waldchisq}
\title{Wald-type chi-squared test}
\usage{
waldchisq(mod, mod_null = NULL, restrictions = NULL,
  restrictions.phi = NULL)
}
\arguments{
\item{mod}{unrestricted model fit from \code{bbdml}}

\item{mod_null}{Optional. Restricted model fit from \code{bbdml}. If not included, need to include \code{restrictions} or \code{restrictions.phi}.}

\item{restrictions}{Optional. Defaults to \code{NULL}. Numeric vector indicating the parameters associated with the abundance to test, or character vector with name of variable to test. Note that \code{1} is the intercept associated with the abundance.}

\item{restrictions.phi}{Optional. Defaults to \code{NULL}. Numeric vector indicating the parameters associated with the dispersion to test, or character vector with name of variable to test. Note that \code{1} is the intercept associated with the dispersion.}
}
\value{
P-value.
}
\description{
Wald-type chi-squared test
}
\examples{
\dontrun{
TODO
}
}
