% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{waldchisq_test}
\alias{waldchisq_test}
\title{Wald-type chi-squared test statistic}
\usage{
waldchisq_test(mod, restrictions = NULL, restrictions.phi = NULL)
}
\arguments{
\item{mod}{an object of class \code{bbdml}}

\item{restrictions}{Optional. Defaults to \code{NULL}. Numeric vector indicating the parameters associated with the abundance to test, or character vector with name of variable to test. Note that \code{1} is the intercept associated with the abundance.}

\item{restrictions.phi}{Optional. Defaults to \code{NULL}. Numeric vector indicating the parameters associated with the dispersion to test, or character vector with name of variable to test. Note that \code{1} is the intercept associated with the dispersion.}
}
\value{
Test statistic for Wald test.
}
\description{
This is a helper function and not intended for users
}
\examples{
\dontrun{
data(soil_phylo)
soil <- soil_phylo \%>\%
phyloseq::subset_samples(DayAmdmt \%in\% c(11,21)) \%>\%
tax_glom("Phylum")
mod1 <- bbdml(formula = OTU.1 ~ DayAmdmt,
phi.formula = ~ DayAmdmt,
data = soil)

waldchisq_test(mod = mod1, restrictions = 2)
}
}
