% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{waldt}
\alias{waldt}
\title{Wald-type t test (model-based or robust)}
\usage{
waldt(mod)
}
\arguments{
\item{mod}{an object of class \code{bbdml}}
}
\value{
Matrix with wald test statistics and p-values. Only performs univariate tests.
}
\description{
Wald-type t test (model-based or robust)
}
\examples{
data(soil_phylum_small_otu1)
mod <- bbdml(formula = cbind(W, M - W) ~ DayAmdmt,
phi.formula = ~ DayAmdmt,
data = soil_phylum_small_otu1)
waldt(mod1)

}
